refactor(activerecord): consolidate DatabaseStatements module into abstract/database-statements.ts
Moves the `DatabaseStatements` module-object export (introduced for
`include(AbstractAdapter, DatabaseStatements)`) into the canonical
Rails-module file at `connection-adapters/abstract/database-statements.ts`
— one home for the Rails DatabaseStatements surface. Deletes the
short-lived `database-statements-defaults.ts`.

The file now contains both: the Rails-module-mirror `export function`
surface (selectAll/isWriteQuery/execute stubs + working exec/delete/
update/insert/transaction/query helpers used as utilities) AND the
`DatabaseStatements` module object mixed onto AbstractAdapter's
prototype via `include()`. The two layers are separated by a header
comment explaining the split.
