+++ packages/activerecord/src/connection-adapters/abstract/database-statements.ts
@@ -11,6 +11,7 @@ import { TransactionIsolationError } from "../../errors.js";
 import { quote, quoteTableName, quoteColumnName } from "./quoting.js";
 import { TransactionManager } from "./transaction.js";
 import { Result } from "../../result.js";
+import { isWriteQuerySql } from "../sql-classification.js";
 
 /**
  * Host interface for DatabaseStatements mixin methods that need adapter context.
@@ -1113,3 +1114,143 @@ export { insert as create };
  * Alias: remove = delete (backwards compat with prior TS API)
  */
 export { deleteStatement as remove };
+
+// ---------------------------------------------------------------------------
+// Module object
+//
+// Rails' `include DatabaseStatements` — applied to AbstractAdapter (and
+// the test SchemaAdapter) via activesupport's `include()` helper. The
+// bodies delegate to the adapter's `execute` / `executeMutation`
+// primitives, matching Rails' pattern where module methods call down
+// into the adapter's `internal_exec_query` / `exec_query` layer.
+//
+// The file-level `export function selectAll` / etc. above are the
+// unbound Rails-module-style references (Rails' default behavior of
+// "must be implemented by subclass"); the module object below is the
+// concrete default set mixed onto AbstractAdapter's prototype.
+// ---------------------------------------------------------------------------
+
+interface DatabaseStatementsMixinHost {
+  execute(sql: string, binds?: unknown[]): Promise<Record<string, unknown>[]>;
+  executeMutation(sql: string, binds?: unknown[]): Promise<number>;
+  execQuery(sql: string, name?: string | null, binds?: unknown[]): Promise<Result>;
+}
+
+export const DatabaseStatements = {
+  async selectAll(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    name?: string | null,
+    binds?: unknown[],
+  ): Promise<Result> {
+    // Rails: select_all → internal_exec_query → exec_query. Delegating
+    // here lets adapters that override execQuery (e.g. PostgreSQLAdapter,
+    // which populates columnTypes via its type_map) have their override
+    // picked up automatically.
+    return this.execQuery(sql, name, binds);
+  },
+
+  async selectOne(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<Record<string, unknown> | undefined> {
+    const rows = await this.execute(sql, binds);
+    return rows[0];
+  },
+
+  async selectValue(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<unknown> {
+    const rows = await this.execute(sql, binds);
+    if (rows.length === 0) return undefined;
+    const keys = Object.keys(rows[0]);
+    return keys.length > 0 ? rows[0][keys[0]] : undefined;
+  },
+
+  async selectValues(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<unknown[]> {
+    const rows = await this.execute(sql, binds);
+    if (rows.length === 0) return [];
+    const firstKey = Object.keys(rows[0])[0];
+    if (firstKey === undefined) return rows.map(() => undefined);
+    return rows.map((row) => row[firstKey]);
+  },
+
+  async selectRows(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<unknown[][]> {
+    const rows = await this.execute(sql, binds);
+    if (rows.length === 0) return [];
+    const keys = Object.keys(rows[0]);
+    return rows.map((row) => keys.map((key) => row[key]));
+  },
+
+  async execQuery(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<Result> {
+    const rows = await this.execute(sql, binds);
+    return Result.fromRowHashes(rows);
+  },
+
+  async execInsert(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<number> {
+    return this.executeMutation(sql, binds);
+  },
+
+  async execDelete(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<number> {
+    return this.executeMutation(sql, binds);
+  },
+
+  async execUpdate(
+    this: DatabaseStatementsMixinHost,
+    sql: string,
+    _name?: string | null,
+    binds?: unknown[],
+  ): Promise<number> {
+    return this.executeMutation(sql, binds);
+  },
+
+  isWriteQuery(sql: string): boolean {
+    return isWriteQuerySql(sql);
+  },
+
+  emptyInsertStatementValue(_pk?: string | null): string {
+    return "DEFAULT VALUES";
+  },
+
+  cacheableQuery(
+    this: unknown,
+    klass: {
+      query?(sql: string): unknown;
+      partialQuery?(parts: unknown): unknown;
+      partialQueryCollector?(): unknown;
+    },
+    arel: unknown,
+  ): [unknown, unknown[]] {
+    return cacheableQuery.call(this as never, klass, arel);
+  },
+};
